[Cosmetic] Use NULL subscript ("" instead of " ") for tableName and columnName when generating M code
Currently Octo generates M programs with references like %ydboctocursor(cursorId,"keys",1," "," ")
to correspond to the output table of a SELECT operation. The last two subscripts are " " to
correspond to a table and column with no name (since the output table has no name).

A " " was originally used instead of a "" to avoid a NULSUBSC error when setting null subscripts in the
^%ydboctocursor global (in the schema database file). But now Octo requires null subscripts for various
other reasons (e.g. ORDER BY). Therefore it is cleaner to just use "" instead of " " for the
generated M code e.g. new references would be like %ydboctocursor(cursorId,"keys",1,"","").