[#531] Remove check that a primary key is not NULL
Previously, for queries of the form

```sql
SELECT select_list FROM table WHERE primary_key IS NOT NULL
```

Octo would generate M plans that look like this:

```m
;; This is a generated file; do not modify. Generated M code corresponds to below SQL query
;; ---------------------------------------------------------
;  SELECT id FROM names WHERE id IS NOT NULL;
;; ---------------------------------------------------------

octoPlan0(cursorId)
    DO octoPlan1(cursorId)
    QUIT

octoPlan1(cursorId)
    NEW %ydboctoz KILL %ydboctocursor(cursorId,"keys",3,"","")
    SET %ydboctocursor(cursorId,"keys",1,"NAMES","ID")=""
    FOR  SET %ydboctocursor(cursorId,"keys",1,"NAMES","ID")=$ORDER(^names(%ydboctocursor(cursorId,"keys",1,"NAMES","ID"))) QUIT:(%ydboctocursor(cursorId,"keys",1,"NAMES","ID")="")  DO
    . NEW %ydboctobool SET %ydboctobool=1
    . SET %ydboctobool='$ZYISSQLNULL(%ydboctocursor(cursorId,"keys",1,"NAMES","ID"))
    . IF %ydboctobool DO
    . . SET:1&$INCREMENT(%ydboctocursor(cursorId,"keys",3,"","")) %ydboctoz=42
    . . SET %ydboctocursor(cursorId,"keys",3,"","",%ydboctocursor(cursorId,"keys",3,"",""))=$$str2mval^%ydboctoplanhelpers(%ydboctocursor(cursorId,"keys",1,"NAMES","ID"))
    QUIT
```

Note the `SET %ydboctobool='$ZYISSQLNULL(...)` in the middle of the hot
loop. After this change, that line is no longer emitted, speeding up
queries slightly.

- Make sure ZYSQLNULL is not present
- Use outref instead of checking grep status

If the exit status of the grep is 0
(i.e. a line containing ZYSQLNULL exists in the .m program),
the ! would convert the exit status to a 1 and would cause bats to
signal a failure because of a non-zero exit status. But what if grep
returns 2 due to some error (say no .m file was found). In that case the
! would cause the exit status to be 0 and run the test fine even though
an error occurred.