[#588][#522] Bumped FMT_PLAN_DEFINITION macros to account for plan formatting changes
As part of the #588 and #522 changes, !800 introduced a new `GetScalarOrArray` M routine, replacing the previous `GetScalar` routine. This entails a change in M plan formats, and so calls for bumping `FMT_PLAN_DEFINITION`. However, this bump did not occur on this MR, leading to `test-auto-upgrade` pipeline failures in `TQG04`. This commit fixes this issue by performing the required bump of this macro.