[#205] Valgrind CMake/BATS Integration
This commit adds support to running either all or a percentage (defaulted to 2%
currently) of all the tests in Octo using Valgrind memcheck tool.

This commit also adds two new pipeline jobs (memcheck-ubuntu and memcheck-rocky)
that run only if the branch name has "valgrind" in it or is a scheduled
pipeline. This means that these new pipelines won't run for any normal merge
requests. The memcheck pipelines have an elevated timeout of 3h; the other
pipelines have a timeout of 60m by default. This change necessitated changing
the setting in the GitLab Octo project CI timeout to be increased to 3h from
the default of 60m. 3h is the maximum for the shared runners we are using.

`USE_VALGRIND_MEMCHECK` and `USE_VALGRIND_MEMCHECK_RANDOM` were two new options
that got added to support this work. The latter uses
`USE_VALGRIND_MEMCHECK_RANDOM_PERCENTAGE`, which is defaulted at 2% currently,
but can be edited if you specify it directly using a CMake define or from the
CMake GUI. The 2% was chosen as that keeps the GitLab pipeline run under 60m.

`USE_VALGRIND_MEMCHECK` and `USE_VALGRIND_MEMCHECK_RANDOM` are mutually
exclusive. You can't set them both at the same time.

Using either CMake option will set the option `DISABLE_INSTALL` to `ON'. The
main reason is expediency - doing the tests with installation enabled or
disabled won't make a difference in the Octo leak list, and makes the code
simpler.

The CMocka tests are simpler, so we will start with these. If either Valgrind
option is set (and in case of the _RANDOM version, if the randomization
criteria are met), run the tests using Valgrind, putting the output in
`valgrind_logs` in the build directory.

The other type of tests are the interactive BATS tests: the trick to getting
tests to run octo/rocto under valgrind is to put bash wrapper scripts that use
`exec` in the `valgrind_wrappers` directory, and if one of the Valgrind options
is turned on, put the wrapper script ahead of the actual binaries in `$PATH` in
the set-up step for the tests (i.e. `init_test()` in `test_helpers.bash.in`).

Also in the same section in `init_test()`, since the randomization of the tests
is difficult to debug properly without logging, two files are created if
`USE_VALGRIND_MEMCHECK_RANDOM` is used:

- `random_number_for_valgrind.notice` - This has the random number generated by
  bash, from 0 to 99.
- `running_valgrind.notice` - empty file to say we are running this test under
  valgrind.

The output of the bash tests is in each bats tests directory as
`$BATS_TEST_NAME-valgrind-memcheck-{process_id}-{increment}.log`. The increment
number is to disambiguate files from the same process (e.g. with threads).

Note that we disable `test_query_generator` test if running with
Valgrind, as it takes too long to run.

Three (3) folders are created to add wrapper, log, and helper scripts for
valgrind errors:

- `valgrind_wrappers`
- `valgrind_helpers`
- `valgrind_logs`

The first folder and its contents are created in `cmake/bats-tests.cmake`, as
that's the logical place that uses it (see above). `valgrind_logs` is created
in `cmake/cmocka.cmake`, as that's where it's used. `valgrind_helpers` is
created in `src/CMakeLists.txt`. If Valgrind checking is turned off after it
has been turned on, the folders are deleted in `src/CMakeLists.txt`.

The helper scripts are for interactive use, and are as follows:

- `valgrind_helpers/valgrind_files.sh` - print a list of all the valgrind
  ouptut files
- `valgrind_helpers/valgrind_error_report.sh` - print a list of all valgrind
  output files that contain errors using a grep pattern in file
  `tests/fixtures/valgrind_error_patterns.grep`. This is not ideal - see the
  discusssion later below on why that approach was chosen.
- `valgrind_helpers/octo` and `valgrind_helpers/rocto` - run octo & rocto
  interactively using valgrind, and launch gdb on the first error. These tools
  are designed to debug the actual problems interactively.

Now a discussion on why a grep pattern was chosen: Initially, the previous work
in the area used this approach. The other approach (which I wanted to take) is
to have valgrind exit the process with a positive exit code. This worked well
when running the code interactively, but I couldn't ever get bats to actually
stop on a positive valgrind exit code. Besides, we have several tests that
check the return code, in which case, returning a positive error code will be
confusing to the tests. In the end, I chose to keep the initial approach, but
that's easily changeable in the future.

Note the use of the `tools/ci/ignored-valgrind-errors.supp` file with the
valgrind scripts. This is a pre-existing file that will be added to in future
commits in this Merge Request.

`tools/ci/build.sh` is the pipeline script. It now accepts `random-memcheck` as
the second argument. If supplied, this activates random valgrind memory tests
using the aforementioned cmake flag `USE_VALGRIND_MEMCHECK_RANDOM`.

Previously, due to issues with CentOS 8/Rocky Linux, we couldn't enable using
Valgrind directly on BATS. We are not doing that anymore, so this is no longer
an issue. The pipeline now runs on both Ubuntu and Rocky Linux.

If `random-memcheck` is requested, we disable installation and always run the
full test suite, not the abbreviated one, and then `ctest` is started with a
timeout of 100 minutes (6000 seconds), in case Valgrind hits one of the tests
that takes too long to run. After `ctest` is done, we run the helper script
`valgrind_helpers/valgrind_error_report.sh` and see if it produces any output.
If it does, we set the `exit_status` to 99.

In the block of code that assembles the files `summary_bats_dirs.txt` and
`passed_bats_dirs.txt`, we check if a test has a memory leak, and if it does,
we say so in `summary_bats_dirs.txt`, and not put it in `passed_bats_dirs.txt`.

At the very end of the script, if we are called with "random-memcheck", and if
we have valgrind failures, we dump them to the screen.