Fix/disable tests that fail with Valgrind
- TCR04: Increase Timeout. Test takes way too long with Valgrind.
- TH00{2,5,7,7}: edit function verify_output in test/test_helpers.bash.in
to change the executable name back after valgrind renames it.
- TMU001: Memory usage increases after each query with Valgrind, so the
test's assumption is no longer valid.
- TPC027: This test fails as it limits the number of files that can be
open at a time to test that we don't leak file descriptors. Adding
Valgrind uses a lot more file descriptors. Rather than editing the test
and reference files to have separate logic for Valgrind, disable this
test as the code paths it exercises are well-exercised throughout the
test framework.
- TII07, TDFT04, TUT004: Disable tests for valgrind, as they turn off
write permissions on the directory, resulting in a failure
because valgrind cannot write to the directory.