Add FREE to errors paths in handle_password_message.
This showed up in valgrind as a memory leak in test TPA01. Another test
for it (CMocka test test_handle_password_message) doesn't leak because
it actually frees the memory, but the regular code paths don't. It seems
the correct thing to do is to free memory in the main rocto source code,
not in the test.

Valgrind signature:

```
==8190== 1,025 bytes in 1 blocks are definitely lost in loss record 13 of 52
==8190==    at 0x483E7C5: malloc (vg_replace_malloc.c:380)
==8190==    by 0x124F55: handle_password_message (handle_password_message.c:58)
==8190==    by 0x1181D4: main (rocto.c:425)
==8190==
```