Ignore upstream issue YDB#817
See issue description and ignore file notes for more details.

Note that this happens only on the debug version of YottaDB.

Fails on tests: TW10, TOJ0{2,3}, TS04, TCF018, TC023, TSD04

The use of Valgrind wildcards for the stack (...) is sadly necessary as
the stack is edited by YottaDB M runtime, so it's not available for
inspection.