[#767] Issue ERR_GROUP_BY_OR_AGGREGATE_FUNCTION error even if GROUP BY in sub query uses outer query columns
As an optimization an outer query column in GroupBy of inner query is removed. In the case where only one column is present it becomes an empty GroupBy. Before this change such a case would work as if no GroupBy was used. But now we set `aggregate_function_or_group_by_specified` of the SqlTableAlias to TRUE indicating that the GroupBy exists. This enables GroupBy validations to happen during query qualification resulting in appropriate error being issued when non grouped columns are used in other clauses.

**Testing**
* TGB15 is added to validate this implementation